// src/components/products/product-grid/ProductGriditem.tsx
'use client';

import { Product } from "@/src/interfaces/product.interface"
import Image from "next/image";
import Link from "next/link";
import { useState } from "react";

interface Props {
  product: Product;
}


export const ProductGriditem = ({ product }: Props) => {

const [displayImage, setDisplayImage] = useState( product.images[0] );


  return (
    <div className="rounded-md overflow-hidden fade-in">

      <Link href={`/product/${product.slug}`}>
        <Image
          src={`/products/${ displayImage }`}
          alt={product.title}
          className="w-full object-cover rounded-2xl"
          width={500}
          height={500}
          onMouseEnter={ () => setDisplayImage( product.images[1])}
          onMouseLeave={ () => setDisplayImage( product.images[0])}
        />
      </Link>

      <div className="p-4 flex flex-col">
        <Link className="hover:text-blue-600"
        href={`/product/${product.slug}`}>
          {product.title}
        </Link>
        <span className="font-bold"> Bs {product.price}</span>
      </div>

    </div>
  )
}
